Update substrate version to "polkadot-v0.9.12" (#180)
* update branch for cargo

* Update

* lint

* update branch for cargo

* Update

* lint

* Changes default runtime executor from wasmi to wasmtime

* Random fixes

* fix after review

* lint for imports

* update branch for cargo

* Update

* lint

* update branch for cargo

* Update

* Random fixes

* Changes default runtime executor from wasmi to wasmtime

* fix after review

* lint for imports

* rebase

* bump spec_version

* Comment values

+++ pallet/src/lib.rs
@@ -61,7 +61,6 @@ pub mod pallet {
     }
 
     #[pallet::event]
-    #[pallet::metadata(T::AccountId = "AccountId")]
     #[pallet::generate_deposit(pub(super) fn deposit_event)]
     pub enum Event<T: Config> {
         ChangeValidators(Vec<T::AccountId>, u32),
@@ -238,6 +237,6 @@ pub mod pallet {
             Self::update_authorities(authorities.as_slice());
         }
 
-        fn on_disabled(_validator_index: usize) {}
+        fn on_disabled(_validator_index: u32) {}
     }
 }
